pyproject.toml
[tool.poetry]
name = "ban-jinro-log"
version = "0.1.0"
description = ""
authors = ["visualpaper"]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = "^0.101.1"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
strawberry-graphql = "^0.203.3"
pymongo = "^4.4.1"
tzdata = "^2023.3"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.1.0"
pytest = "^7.4.0"
pytest-mock = "^3.11.1"
pytest-cov = "^4.1.0"
taskipy = "^1.12.0"
isort = "^5.12.0"
autoflake = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[tool.flake8]
max-line-length = 88

[tool.autoflake]
imports = ["fastapi", "motor"]

[tool.pytest.ini_options]
# tests が src と同階層にいる場合、tests から src を参照するために必要
pythonpath = [
  ".", "src"
]

[tool.coverage.report]
# pytest カバレッジ用の設定
## show_missing: test がない行を表示する (結果が長くなるのでデフォルトは false としています)
## skip_empty: 空ファイルを除外する (主に __init__.py 除外目的です)
show_missing = false
skip_empty = true

[tool.taskipy.tasks]
# skip-string-normalization で ' を " に変えることを抑制しています。
# ※ PEP8 では規定されていないが好みだと思うので抑制しています。
format = """\
         autoflake -r src tests --remove-all-unused-imports --in-place\
         && isort src tests\
         && black src tests --skip-string-normalization
         """

# max-line-length で black の定義 88 に合わせています。
# (参照) https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html#line-length
#
# W503 と E203 は black の定義と合わせています。
# (参照) https://rcmdnk.com/blog/2019/11/04/computer-python/
#        https://org-technology.com/posts/python-black.html
lint = "flake8 src tests --ignore=E203,W503 --max-line-length=88"

# -s: テスト実行中に print 文の出力を標準出力に書き出す。
# -v: テストごとの Success/Failed を表示する。
# --cov: カバレッジを表示する。
test = "pytest tests -s -vv --cov"

# Local 実行用
# ※ poetry shell 後に実行可能な点に注意してください。
start = "uvicorn src.main:app --reload"
